Skip tests for indices failing on Neo4j 5
Neo4j 5 removed a number of deprecated procedures used by Neo4p to implement indices. As a result, indices don't work with Neo4p 0.4003 on Neo4j 5.

Simply trying to replace the deprecated calls in DriverActions.pm with the new syntax seemed to work in Neo4j 4, but not Neo4 5. The issue is possibly related to the way Neo4p uses properties on the index. There might be changes that I missed in Neo4j 5 in addition to the new syntax.

But since Neo4j 5 currently isn't officially supported by Neo4p, it might not matter much. This change for now simply skips the failing tests.

Note that Neo4j < 5.17 doesn't allow query parameters in the new index syntax.