Fix dependencies
REST::Neo4p 0.4000+ requires HTTP::Tiny, which was first released with Perl v5.13.9. This change adds it to the required modules for compatibility with earlier versions of Perl.

Mojo::Exception was only used in a single test, and that test is currently skipped by default. So there's currently no reason to force users to install Mojolicious.